% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/calc_sum.R
\name{calculate_sum}
\alias{calculate_sum}
\alias{calculateSum}
\title{calculate_sum}
\usage{
calculateSum(data, processors)

calculate_sum(data, processors = parallelly::availableCores())
}
\arguments{
\item{data}{vector, containing Integers to add}

\item{processors}{Integer, number of cores to use. Default = all available}
}
\value{
Integer, containing sum of data

Integer, containing sum of data
}
\description{
Calculate the sum of integers

Calculate the sum of integers
}
\examples{

data <- 1:20
total <- calculate_sum(data)

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
